% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{chullcheck}
\alias{chullcheck}
\title{Checks whether the point test is in the convex hull of the rows of X.}
\usage{
chullcheck(X, test)
}
\arguments{
\item{X}{An n by d matrix.}

\item{test}{An n by 1 vector.}
}
\value{
A binary variable, which is 1 if the point test is in the
                      convex hull of the rows of X, and 0 otherwise.
}
\description{
This function will take the matrix X and the vector test as inputs,
and return a binary variable, which is 1 if the point test is in the
convex hull of the rows of X, and 0 otherwise.
}
\examples{
\dontrun{
## We demonstrate an example, where the point test is in the convex hull of the rows of X.
X<-matrix(c(0,0,1,1,0,1,0,1),4,2)
test<-c(0.3,0.4)

out = chullcheck(X,test)
}
}
