% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{chulldowncheck}
\alias{chulldowncheck}
\title{Checks whether the point test is in the lower orthant of the convex
hull of the rows of X.}
\usage{
chulldowncheck(X, test)
}
\arguments{
\item{X}{An n by d matrix.}

\item{test}{An n by 1 vector.}
}
\value{
A binary variable, which is 1 if the point test
                      is in the lower orthant of the convex hull of the
                      rows of X, and 0 otherwise.
}
\description{
Checks whether the point test is in the lower orthant of the convex
hull of the rows of X.
}
\details{
This function will take the matrix X and the
vector test as inputs, and return a binary variable, which is 1 if
the point test is in the lower orthant of the convex hull of the
rows of X, and 0 otherwise.
}
\examples{
\dontrun{
## We demonstrate an example, where the point test is in the lower orthant of the convex hull
## of the rows of X.
X<-matrix(c(0,0,1,1,0,1,0,1),4,2)
test<-c(-1,-2)

out = chulldowncheck(X,test)
}
}
