% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdditionalFucntions.R
\name{interpol.func}
\alias{interpol.func}
\title{Interpolates function under shape and/or monotocity constraints}
\usage{
interpol.func(
  X,
  cplex.fhat,
  nondatapoint,
  Shape = c("QuasiConvex", "none"),
  Monotone = c("non.inc", "non.dec", "none")
)
}
\arguments{
\item{X}{An n by d matrix of data points.}

\item{cplex.fhat}{An n by 1 vector of functional values evaluated at the rows of X.}

\item{nondatapoint}{A d by 1 non-data point where the function needs to be interpolated/extrapolated.}

\item{Shape}{A categorical variable indicating the shape of the function.
The user must input one of the following two types:
"QuasiConvex": means quasiconvex function,
"none": means no shape restriction is imposed.}

\item{Monotone}{A categorical variable indicating the monotonicity pattern of the function.
The user must input one of the following three types:
"non.inc": means nonincreasing function,
"non.dec": means nondecreasing function,
"none": means no monotonicity restriction is imposed.}
}
\value{
The interpolated/extrapolated value of the function (subject to
                      the shape and monotonicity constraints specified) at the non-data point.
}
\description{
Interpolates/extrapolates a quasiconvex function with/without additional
monotonicity constraints (already defined on some data points), to a non-data point.
This function takes a matrix of data points, a vector of the corresponding functional
values at the data points, a non-data point, and shape and monotonicity specifications
as inputs, and returns the interpolated/extrapolated value of the function (subject to
the shape and monotonicity constraints specified) at the non-data point.
}
