% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{interpolation.matrix.func}
\alias{interpolation.matrix.func}
\title{Interpolates function under shape and/or monotocity constraints}
\usage{
interpolation.matrix.func(
  X,
  fin,
  nondatapoint_matrix,
  Shape = c("QuasiConvex", "QuasiConcave", "none"),
  Monotone = c("non.inc", "non.dec", "none")
)
}
\arguments{
\item{X}{An n by d matrix of regressors.}

\item{fin}{An n by 1 vector of fitted values.}

\item{nondatapoint_matrix}{An N by d matrix whose rows may not be among the rows of X.}

\item{Shape}{A categorical variable indicating the type of regression.
 The user must input one of the following three types:
 "QuasiConvex": means QuasiConvex regression -- with possible monotonicty constraint enforced by 'Monontone'
 "QuasiConcave": means QuasiConcave regression -- with possible monotonicty constraint enforced by 'Monontone'
"none": means there is neither}

\item{Monotone}{A categorical variable denoting the type of regression in terms of monotonicity, in addition to possible shape constraint imposed by the 'Shape' parameter
The user must input one of the following three types:
"non.inc": means nonincreasing
"non.dec": means nondecreasing
"non": means quasiconvex regression.}
}
\value{
A vector, representing a quasiconvex and monotone (type of monotonicity
                                 determined by the user entered option for the input variable Monotone)
                                 function evaluated at the rows of the matrix nondatapoint_matrix,
                                 which is an extrapolation of the fitted values in the vector fin.
}
\description{
Assumes that 'fin' satisfy some monotonicity and shape constraint
fit obtained from the matrix X of regressors and some response vector.
This function will take as inputs, the matrix X, the fitted vector fin,
a matrix nondatapoint_matrix whose rows may not be among those of X, and a
categorical variable Shape which denotes the type of regression in terms of
shape, and Monotone which denotes the type of regression in terms of
monotonicity. It will return a quasiconvex/quasiconcave/none and monotone (type of
monotonicity determined by the user entered option for the variable Monotone) function
evaluated at the rows of the matrix nondatapoint_matrix, which is an
extrapolation of the fitted values in the vector fin. This function is user friendly wrapper for 'interpol.func'.
}
