% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{qconvcheck}
\alias{qconvcheck}
\title{Checks whether the points in the vector y are actually the outputs of a single quasiconvex function applied to the corresponding rows of the matrix X.}
\usage{
qconvcheck(X, y, tol = 0.001)
}
\arguments{
\item{X}{An n by d matrix.}

\item{y}{An n by 1 vector.}

\item{tol}{A small positive quantity.}
}
\value{
A string, indicating whether the vector y is, or not
                      a quasiconvex realization of X.
}
\description{
This function will take the matrix X and the vector y as inputs, and return a string,
indicating whether the vector y is or not a quasiconvex realization of X.
}
\examples{
\dontrun{
library(MASS)
## We demonstrate an example, where the regressor matrix X is Gaussian, and the response vector y
## is a real-valued quasiconvex function applied to the rows of X.
X = mvrnorm(20,rep(0,8), diag(8), tol=1e-06,empirical=FALSE)
y = exp(-rowSums(X))

out = qconvcheck(X,y, tol = 1e-3)
}
}
