% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{Quasi.reg}
\alias{Quasi.reg}
\title{Computes quasiconvex/quasiconcave least squares estimate with or without additional monotonicity constraint via CPLEX  or Gurobi.}
\usage{
Quasi.reg(
  X,
  y,
  Shape = c("QuasiConvex", "QuasiConcave"),
  Monotone = c("non.inc", "non.dec", "none"),
  Max.b = 10^4,
  MM = 10^4,
  ep = 0.001,
  time.limit = 200,
  tol = 1e-06,
  Init.start = NULL,
  optimizer = c("cplex", "gurobi")
)
}
\arguments{
\item{X}{An n by d matrix of regressors.}

\item{y}{An n by 1 vector of responses.}

\item{Shape}{A categorical variable indicating the type of regression.
The user must input one of the following three types:
"QuasiConvex": means QuasiConvex regression -- with possible monotonicty constraint enforced by 'Monontone'
"QuasiConcave": means QuasiConcave regression -- with possible monotonicty constraint enforced by 'Monontone'}

\item{Monotone}{A categorical variable indicating the type of  monotonicitty constraint on top of the shape constraint enforced by 'Shape' parameter
The user must input one of the following three types:
"non.inc": means additional nonincreasing.
"non.dec": means nondecreasing.
"non": means no additional monotonicity constraint}

\item{Max.b}{Bound on the absolute values of the variables.}

\item{MM}{The 'M' parameter in the big-M method applied by us.}

\item{ep}{A small positive quantity to convert open constraints to close constraints.}

\item{time.limit}{Time limit in seconds of call to optimizer. Can be any nonnegative number.}

\item{tol}{Relative MIP optimality gap tolerance. Can be any nonnegative number. Default 1e-4.}

\item{Init.start}{A matrix for warm start. Default is NULL. If provided should be matrix with n rows.}

\item{optimizer}{Which optimizer to use when solving the MIQP.  The user must input one of the following two types: 'gurobi' or 'cplex'. Make sure they installed locally and accesbible frome the R envoirment.}
}
\value{
An object of class 'Quasi.reg' that contain among other things, the
                      least squares estimates at the regressors (within f.hat) and
                       the optimal objective function (within minvalue), list of solutions based on each of the intial starts (within FhatList), the initial start that lead to the the optimum (within K), and the corresponding convergence status (within status)
}
\description{
This function will take in the multivariate regressors and the response, and return the least squares estimates at the regressors, based on the response. X,y and Monotone must be supplied, though all other parameters have pre-set values the user can proceed with unless they wish to change the prior specification.
}
\examples{
n=20
d=4
library(MASS)
## We demonstrate an example, where the regressor matrix X is Gaussian, and the response vector y
## is a noisy version of a real-valued quasiconvex, decreasing function applied to the rows of X.
X = matrix(runif(n*d, 0, 3), nrow=n)
y = exp(-rowSums(X)) + rnorm(n)
ret = Quasi.reg(X, y,Shape ="QuasiConvex", Monotone = "non.inc",tol =  1e-06, optimizer= "gurobi")

}
