% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdditionalFucntions.R
\name{QuasiConv.control}
\alias{QuasiConv.control}
\title{Creates matrices, vectors and variable bounds corresponding to the objective and the constraints
of the mixed integer quadratic optimization.}
\usage{
QuasiConv.control(
  X,
  y,
  Max.b,
  MM,
  ep,
  Monotone = c("non.inc", "non.dec", "none")
)
}
\arguments{
\item{X}{An n by d matrix of data points.}

\item{y}{An n by 1 vector of responses.}

\item{Max.b}{Bound on the absolute values of the variables.}

\item{MM}{The 'M' parameter in the big-M method applied by us.}

\item{ep}{A small positive quantity to convert open constraints to close constraints.}

\item{Monotone}{A categorical variable indicating the monotonicity pattern of the function.
The user must input one of the following three types:
"non.inc": means nonincreasing function,
"non.dec": means nondecreasing function,
"none": means no monotonicity restriction is imposed.}
}
\value{
List of matrices, vectors and variable bounds corresponding to the objective function
                      and the constraints of the mixed integer quadratic optimization. The output of this
                      function can be used by CPLEX/GUROBI for implementing the mixed integer optimization.
}
\description{
This function takes the matrix of regressors, the response
variables, the monotonicity specification and some other bounds and tuning parameters as inputs, and returns
a list of matrices, vectors and variable bounds corresponding to the objective function and the constraints
of the mixed integer quadratic optimization. The output of this function can be used by CPLEX/GUROBI for
implementing the mixed integer optimization.
}
